// InventoryManagementUI/src/openapi/services/model/users.ts
/**
 * Seyir Füzeleri Sistem Mühendisliği Envanter Uygulaması API
 * NodeJS - ExpressJS API Docs / Fevzi TEMİZ
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Employees } from './employees';
import { Roles } from './roles';


export interface Users { 
    id: number;
    userName: string;
    isActive: boolean;
    role: Roles;
    employee: Employees;
}